refactor app folders, move book list and book details to screens
